% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmycSimulatedTree.R
\name{gmycSimulatedTree}
\alias{gmycSimulatedTree}
\title{Simulate tree for GMYC}
\usage{
gmycSimulatedTree(SpeciationTree, SpeciesResult)
}
\arguments{
\item{SpeciationTree}{tree}

\item{SpeciesResult}{a data.frame with an index for species in the first column,
the number of individuals per species in the 2nd column
and the nucleotide diversity (as proxy for population size) in the 3rd column
(NA for species with one individual)}

\item{Scale}{should tree be scaled to unit deepth? (default TRUE)}
}
\description{
This function is a fork of GMYC.simulated.tree from the P2C2M.GMYC package
and graft the speciation and coalescent tree and correct the branch lengths.
}
\author{
Torsten Hauffe (with code forked from Emanuel M. Fonseca and Drew J. Duckett)
}
