% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.traitgmyc.R
\name{plot.traitgmyc}
\alias{plot.traitgmyc}
\title{Plot the result of traitgmyc}
\usage{
\method{plot}{traitgmyc}(x, ask = TRUE, ...)
}
\arguments{
\item{x}{An object of class traitgmyc.}

\item{ask}{logical. If TRUE (and the R session is interactive) the user is asked for input, before a new figure is drawn.}

\item{...}{Other parameters to be passed through to plotting functions.}
}
\value{
A plot with three panels.
}
\description{
This function... .
}
\examples{
1 + 2
# plot()
}
\author{
Torsten Hauffe and Robin Ackermann
}
