% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSigma.R
\name{simSigma}
\alias{simSigma}
\title{Simulate variance-covariance matrix for morphological evolution}
\usage{
simSigma(Ntraits, Cor = NULL, Sigma2 = NULL)
}
\arguments{
\item{Ntraits}{number of traits}

\item{Cor}{correlation between traits. Default between -1 and 1.
Optional, can be fixed to be equal between all traits by giving one value or
Ntraits*(Ntraits-1)/2}

\item{Sigma2}{Brownian motion rate. Default between 1e-4 and 0.2
Optional, can be fixed to be equal for all traits or one value per trait}
}
\value{
matrix Ntrait x Ntrait for simulating trait evolution
}
\description{
This function generates a variance-covariance matrix for
morphological evolution, defining the rate at which traits evolve and their
evolutionary correlation.
}
\author{
Torsten Hauffe
}
