% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTraitsIndividuals.R
\name{simTraitsIndividuals}
\alias{simTraitsIndividuals}
\title{Simulate traits}
\usage{
simTraitsIndividuals(
  Tree,
  Ntraits = 1,
  IndPop,
  PopFluc = NULL,
  GenLength = 1,
  Sigma2 = NULL,
  Cor = NULL,
  Alpha = NULL,
  Scenario = "A",
  TraitModel = "BM"
)
}
\arguments{
\item{Tree}{The species tree}

\item{Ntraits}{Number of traits to simulate}

\item{IndPop}{Matrix or data.frame with the number of individuals
per species in the first column and the species' population size in the second column.
This element is returned by simGenealogy ('Species').}

\item{PopFluc}{Optional matrix or data.frame with the parameters for a fluctuating population size
(returned as third element form simGenealogy)}

\item{Sigma2}{Optional Brownian motion rate for trait evolution.
Vector of length one (the same for all traits and species) or a seperate rate per species.}

\item{Cor}{Optional correlation strength between traits in the range \link{-1,1}.
Either a vector of length one (same strength between all traits) or Ntraits*(Ntraits-1)/2.}

\item{Alpha}{Optional strength of pull towards the morphological optima.
Only effective for scenarios A, B and C.
If not specified, optima is (theoretically) reached after half of the tree depth.}

\item{TraitModel}{Model of trait evolution. Ether Brownian motion (BM)
or Orenstein-Uhlenbeck (OU).}

\item{GenLen}{Optional vector of length one for generation length (the same for all species)}
}
\value{
A matrix with traits in columns and names of individuals per species as rownames
}
\description{
This function simulates traits for a given number of
individuals per species and the population size for the species
}
\references{
Fujisawa, T. and T. Barraclough (2013): Delimiting species using
single-locus data and the Generalized Mixed Yule Coalescent approach:
a revised method and evaluation on simulated data sets.
Systematic Biology 62(5), 707-724.
}
\author{
Torsten Hauffe
}
