\name{TraitGMYC-package}
\alias{TraitGMYC-package}
\alias{TraitGMYC}
\docType{package}
\title{
\packageTitle{TraitGMYC}
}
\description{
\packageDescription{TraitGMYC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TraitGMYC}
\packageIndices{TraitGMYC}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{TraitGMYC}

Maintainer: \packageMaintainer{TraitGMYC}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
