% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yuleTree.R
\name{yuleTree}
\alias{yuleTree}
\title{Simulate tree for GMYC}
\usage{
gmycSimulatedTree(SpeciationTree, EmpiricalTree)
}
\arguments{
\item{SpeciesResult}{a data.frame with an index for species in the first column,
the number of individuals per species in the 2nd column
and the nucleotide diversity (as proxy for population size) in the 3rd column
(NA for species with one individual)}

\item{EmpiricalTree}{Empirical tree used for species delimitation}

\item{Scale}{should tree be scaled to unit depth? (default TRUE)}
}
\description{
This function is a fork of GMYC.simulated.tree from the P2C2M.GMYC package
It simulates a tree under the yule (pure-birth) process with the same number of species
that was delimited using the empirical dataset
}
\author{
Torsten Hauffe (with code forked from Emanuel M. Fonseca and Drew J. Duckett)
}
