% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simGenealogy.R
\name{simGenealogy}
\alias{simGenealogy}
\title{Simulate genealogies}
\usage{
simGenealogy(
  Species = 30,
  Scenario = "A",
  Ind = 5,
  PopSize = 10000,
  Nsim = 1,
  Scale = NULL
)
}
\arguments{
\item{Species}{The species tree within the gene trees should be simulated
for Scenarios B-G. In case of Scenario A the number of species.
If Nsim > 1 there should be multiphylo}

\item{Scenario}{Scenario of Fujisawa and Barraclough (2013)
\itemize{
\item A Null model assuming a neutral coalescent process in a single population
\item B Diversification (coalescence within a species tree)
\item D Fluctuating population sizes
}}

\item{Ind}{Individuals per species.
Either a vector of length 1 (i.e. same number of individuals per species) or
of the same length as Ntip(SpeciesTree)}

\item{PopSize}{Population size \ifelse{html}{\out{<i>N</i><sub>e</sub>}}{\eqn{Ne}}.
Assumed to be the same for each species but can vary over time in scenario D.
Either a vector of length 1 or of the same length as Nsim.}

\item{Nsim}{Number of simulated genealogies}

\item{Scale}{Scale depth of genealogy to this value (e.g. 1)}
}
\value{
A genealogy, which for scenarios B-G is simulated
within the provided species tree.
}
\description{
This function simulates genealogies
using the multispecies coalescence approach of Fujisawa and Barraclough (2013).
}
\details{
Scenarios C1 and C2 of incomplete taxon sampling or
non-zero rate of species extinction from Fujisawa and Barraclough (2013) are
possible to simulate by using a different species tree (see examples).
Moreover, the species tree could be also generated by the protracted birth-death
process to mirror the genealogy simulation of Sukumaran and Knowles (2017).
}
\examples{
# Scenario A
GeneTree <- simGenealogy(Species = 30,
                         Scenario = "A",
                         Ind = 5,
                         PopSize = 10000)
plot(GeneTree, cex = 0.4)
# Scenario B
SpeciesTree <- pbtree(b = 0.27, n = 30)
GeneTree <- simGenealogy(Species = SpeciesTree,
                         Scenario = "B",
                         Ind = 5,
                         PopSize = 10000)
plot(GeneTree, cex = 0.4)

# Scenario C1
FullTree <- pbtree(b = 0.27, n = 50)
SpeciesTree <- drop.tip(FullTree, sample(FullTree$tip.labels, 20))
GeneTree <- simGenealogy(Species = SpeciesTree,
                         Scenario = "C1",
                         Ind = 5,
                         PopSize = 10000)
plot(GeneTree, cex = 0.4)
# Scenario C2
}
\references{
Fujisawa, T. and T. Barraclough (2013): Delimiting species using
single-locus data and the Generalized Mixed Yule Coalescent approach:
a revised method and evaluation on simulated data sets.
Systematic Biology 62(5), 707-724.
}
\author{
Torsten Hauffe but most of the code was forked
from Fujisawa and Barraclough (2013)
}
